+++ src/bot/bot_factory.py
@@ -1,11 +1,14 @@
 import os
-from telegram.ext import Updater, CommandHandler
+from telegram.ext import Updater, CommandHandler, MessageHandler, Filters
 
 from src.bot import logger_factory
 from src.bot.commands import health_check_command
 from src.bot.commands import user_commands
 from src.bot.commands import help_command
 from src.bot.commands import match_commands
+
+from src.bot.message_handlers.freedom_units import convert_to_freedom_units
+
 from src.constants import LOG_LEVEL_MAP
 
 
@@ -24,5 +27,9 @@ def create_bot():
     dp.add_handler(CommandHandler("help", help_command.run_help_command))
     dp.add_handler(CommandHandler("lastmatch", match_commands.run_last_match_command))
     dp.add_handler(CommandHandler("match", match_commands.run_get_match_by_match_id))
+
+    dp.add_handler(
+        MessageHandler(Filters.text & ~Filters.command, convert_to_freedom_units)
+    )
 
     return updater, logger